chore: move typography to ui folder
